+++ package.json
@@ -26,8 +26,8 @@
     "iframe-resizer": "^4.3.11",
     "lodash.chunk": "^4.2.0",
     "polished": "^4.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-error-boundary": "^3.1.4",
     "react-hook-form": "^7.37.0",
     "react-query": "^3.39.2",
@@ -43,9 +43,9 @@
     "@testing-library/react": "^12.1.2",
     "@types/jest": "^27.4.0",
     "@types/lodash.chunk": "^4.2.7",
-    "@types/react": "^17.0.33",
+    "@types/react": "^18.3.3",
     "@types/react-copy-to-clipboard": "^5.0.4",
-    "@types/react-dom": "^17.0.10",
+    "@types/react-dom": "^18.3.0",
     "@types/styled-components": "^5.1.23",
     "@types/testing-library__jest-dom": "^5.14.2",
     "@types/uuid": "^8.3.4",
