+++ src/App.test.tsx
@@ -1,5 +1,5 @@
 import { render, waitFor } from "@testing-library/react";
-import App from "./App";
+import { App } from "./App";
 
 test("renders App component", async () => {
     const { getByText } = render(<App />);
