+++ src/components/AzureIcon.tsx
