+++ src/components/common/Label/index.ts
@@ -1,0 +1,1 @@
+export { Label } from "./Label";
