+++ src/components/common/Layout/Container.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  min-height: 2em;
+  margin: 8px;
+`;
+
+export { Container };
