+++ src/main.css
@@ -1,7 +1,3 @@
-body {
-  margin: 10px 5px 5px 5px;
-}
-
 input::-webkit-outer-spin-button,
 input::-webkit-inner-spin-button {
   appearance: none;
