+++ src/pages/Admin/AccountNamePatToken.tsx
@@ -5,7 +5,7 @@ import {
   useInitialisedDeskproAppClient,
 } from "@deskpro/app-sdk";
 import { useEffect, useState } from "react";
-import { Input } from "../../Components/Input";
+import { Input } from "../../components/Input";
 import { Settings } from "../../types";
 import { preInstallGetCurrentUserPremise } from "../../api/preInstallApi";
