+++ src/pages/FindOrCreateItems.tsx
@@ -1,17 +1,16 @@
-import { Stack } from "@deskpro/deskpro-ui";
-import { TwoButtonGroup } from "@deskpro/app-sdk";
 import { useState } from "react";
+import { TwoButtonGroup } from "@deskpro/app-sdk";
 import { faMagnifyingGlass, faPlus } from "@fortawesome/free-solid-svg-icons";
-
-import { CreateItem } from "../Components/Items/CreateItem";
-import { FindItem } from "../Components/Items/FindItem";
+import { Container } from "../components/common";
+import { CreateItem } from "../components/Items/CreateItem";
+import { FindItem } from "../components/Items/FindItem";
 
 export const FindOrCreateItems = () => {
   const [page, setPage] = useState<0 | 1>(0);
 
   return (
-    <Stack vertical>
-      <Stack style={{ alignSelf: "center" }}>
+    <>
+      <Container>
         <TwoButtonGroup
           selected={
             {
@@ -21,19 +20,18 @@ export const FindOrCreateItems = () => {
           }
           oneIcon={faMagnifyingGlass}
           twoIcon={faPlus}
-          oneLabel="Find Item⠀⠀"
-          twoLabel="Create Item⠀⠀"
+          oneLabel="Find Item"
+          twoLabel="Create Item"
           oneOnClick={() => setPage(0)}
           twoOnClick={() => setPage(1)}
-        ></TwoButtonGroup>
-      </Stack>
-
+        />
+      </Container>
       {
         {
           0: <FindItem />,
           1: <CreateItem />,
         }[page]
       }
-    </Stack>
+    </>
   );
 };
